make utils file
